requirements.txt
streamlit
pdfplumber
python-docx
sentence-transformers
scikit-learn